refactor(map): wave 4 render split, marker types, and CI ratchets (#4112)
Split near-cap render modules by concern, extract marker view types and
generator arg/test helpers, deduplicate legend height via
legendHeightForLineCount, and add repo.map_* headroom gates.
