+++ SPEC/program/repo-lint.md
@@ -77,6 +77,7 @@
 | `tool/check_map_gen_stage_protocol.dart` | Require the shared `MapGenStage` contract in `packages/colonizethis_map/lib/src/gen/map_gen_stage.dart` and `implements MapGenStage` (or the uniform `MapGenPass`) on each bound generator service family (`TileMapGenLandSeeds`, `TileMapGenLakesProvinces`, `ContinentJoinPass`, `TerrainJitterPass`, `SeaZoneSubdividePass`, `TileMapGenTerrainResource`) so pass orchestration shares a uniform params + grid-in/grid-out protocol, and require **at least 4** bound families to adopt the uniform `MapGenPass` entry point (raised from 3 after the former join-sea family was split into three standalone `MapGenPass` passes, removing the last `MapGenStage`-only exemption) (Refs #3459, #3574, #3588); rule `repo.map_gen_stage_protocol` |
 | `tool/check_map_no_partfile_classes.dart` | Two map structure gates share this script. `repo.map_no_partfile_classes` forbids any `part of` file under `packages/colonizethis_map/lib/src/**` from declaring a top-level non-private class. `repo.map_gen_no_new_partfiles` forbids **any** `part` / `part of` directive under the generation layer (`lib/src/gen/**`) and view layer (`lib/src/view/**`) with an **empty** allowlist after #4022 promoted `tile_map_generator_types.dart` to a standalone library (closing the temporary #3588 exception); rules `repo.map_no_partfile_classes`, `repo.map_gen_no_new_partfiles` |
 | `tool/check_map_lib_file_size.dart` | Enforce that every `.dart` file under the colonizethis_map generation layer (`lib/src/gen/**`), view layer (`lib/src/view/**`), and render layer (`lib/src/render/**`) is **500 non-comment lines or fewer** (fail when strictly greater), reusing `countNonCommentLinesFromSource`; the gate also fails when any scanned root is missing so it cannot silently rot if those layers move. Complements the repository-wide `repo.dart_file_non_comment_line_size` cap so the #3588 / #3846 part-file decomposition keeps each pass/concern small (Refs #3588, #3846); rule `repo.map_lib_file_size` |
+| `tool/check_map_lib_file_size_headroom.dart` | Wave-4 headroom ratchet: every `.dart` file under the same gen/view/render scan roots as `repo.map_lib_file_size` must be **400 non-comment lines or fewer** unless listed in the manifest grandfather baseline (shrink-only allowlist). Complements the 500-NCL hard cap without lowering it (Refs #4112); rule `repo.map_lib_file_size_headroom` |
 | `tool/check_data_lib_file_size.dart` | Enforce that every hand-written `.dart` file under `packages/colonizethis_data/lib/src/**` is **500 non-comment lines or fewer** (fail when strictly greater), reusing `countNonCommentLinesFromSource`; skips generated suffixes and `tech_effect_summary_embed.dart`. Complements the repository-wide 1000 NCL cap so the #4072 topic splits stay small (Refs #4072); rule `repo.data_lib_file_size` |
 | `tool/check_data_victory_config_registry_parity.dart` | Enforce exact set parity between every `const int` / `const double` `k*` in `packages/colonizethis_data/lib/src/ai_victory_config*.dart` and every `victoryConfigIntParam` / `victoryConfigDoubleParam` name string across `ai_parameter_victory_config_params*.dart`. Structured `Map` / `String` constants stay out of scope (Refs #4072); rule `repo.data_victory_config_registry_parity` |
 | `tool/check_map_public_barrel_surface.dart` | Keep the public `colonizethis_map` barrel `packages/colonizethis_map/lib/colonizethis_map.dart` narrowed to intentional entry points: forbid any `export` of the internal-only generation primitives `src/gen/grid_voronoi.dart` (deterministicNoise / assignCellsToNearestSeed), `src/gen/topology_inference.dart` (inferTopologyFromTileMap), and `src/gen/tile_map_grid_graph.dart` (TileMapGridGraph). Same-package tests import these from `src/` directly; comment lines and `import` directives are ignored (Refs #3459 AC4); rule `repo.map_public_barrel_surface` |
@@ -151,6 +152,8 @@
 | `tool/check_map_test_shared_topology_fixtures.dart` | Forbid inline `\bTopologyNode\(` in render/visualizer tests (`test/tile_map_visualization_*`, `test/game_world_state_map_visualizer_test.dart`, `test/multi_region_map_rendering_test.dart`) unless allowlisted: hardcoded exempt files (e.g. `tile_map_topology_validation_test.dart`), `*_validation*_test.dart` basename glob, or a per-site `// map-topology-fixture-exempt: <reason>` marker on the line immediately preceding the inline constructor; `test/support/**` and `test/*_fixtures.dart` are out of scope (Refs #3846); rule `repo.map_test_shared_topology_fixtures` |
 | `tool/check_map_test_run_generation_harness.dart` | Forbid inline `\bTileMapGenerator\s*\(` under `packages/colonizethis_map/test/tile_map_generator_*_test.dart`, `tile_map_generation_*`, `tile_map_pass4_*`, `tile_map_forest_*`, and `tile_map_terrain_*` suites unless the preceding line carries `// map-generation-harness-exempt: <reason>` (constructor/DI probes). Integration generates must use `runTileMapGeneration` from `test/support/tile_map_gen_fixtures.dart`; that support file is out of scope (Refs #4022); rule `repo.map_test_run_generation_harness` |
 | `tool/check_map_test_minimal_game_shared.dart` | Forbid inline `\bGame\(` in the scoped colour/capital/region-input/format/port suites (`faction_ownership_color_test.dart`, `tile_map_capital_markers_test.dart`, `region_map_view_inputs_test.dart`, `map_format_util_test.dart`, `port_icon_placement_test.dart`); those suites must use `minimalGame` from `test/support/init_game_map_view_fixtures.dart`. Full-line comments are stripped; `minimalGame(` is not matched (Refs #4022); rule `repo.map_test_minimal_game_shared` |
+| `tool/check_map_test_file_size.dart` | Physical-line ceiling **400** for every `packages/colonizethis_map/test/**` Dart file unless manifest-baseline grandfathered (Refs #4112); rule `repo.map_test_file_size` |
+| `tool/check_map_render_legend_layout_dedup.dart` | Forbid inline `legendPadding * 2 + legendLines * legendLineHeight` outside `tile_map_visualization_legend_layout.dart` (canonical `legendHeightForLineCount` owner); full-line comments stripped (Refs #4112); rule `repo.map_render_legend_layout_dedup` |
 | `tool/check_combat_test_mirrors_lib.dart` | Keep the `packages/colonizethis_combat/test/**` tree mirrored to `lib/` per the testing rule (`test/` mirrors `lib/`): any Dart source that sits **directly** at the `test/` root (its repo-relative path after `packages/colonizethis_combat/test/` contains no further `/`) is a violation — both `*_test.dart` files and co-located support helpers must live in a subdirectory such as `test/combat/` that mirrors `lib/src/combat/`. Files already nested under a subdirectory (e.g. `test/combat/combat_rng_test.dart`) are compliant; generated suffixes and fixture trees are skipped via the shared scan contract. Complements the testing-rule `test/`-mirrors-`lib/` policy (Refs #3716); rule `repo.combat_test_mirrors_lib` |
 | `tool/check_combat_test_no_local_support.dart` | Forbid package-local `*_test_support.dart` files under `packages/colonizethis_combat/test/**`: canonical home is `colonizethis_combat_test_support` (Refs #3865); rule `repo.combat_test_no_local_support` |
 | `tool/check_combat_test_core_fixtures_shared.dart` | Forbid inline `Game(` construction under `packages/colonizethis_combat/test/**`; use shared builders from `colonizethis_combat_test_support` (Refs #3865); rule `repo.combat_test_core_fixtures_shared` |
